requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
aiosmtplib==2.0.0
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
black==22.12.0
blinker==1.5
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptocode==0.1
cryptography==39.0.0
dnspython==2.2.1
elastic-transport==8.4.0
elasticsearch==8.6.0
email-validator==1.3.0
fastapi==0.88.0
fastapi-mail==1.2.4
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mongoengine==0.25.0
multidict==6.0.4
mypy-extensions==0.4.3
pathspec==0.10.3
Pillow==9.4.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
pycparser==2.21
pycryptodomex==3.16.0
pydantic==1.10.2
PyJWT==2.6.0
pymongo==4.3.3
python-decouple==3.6
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.7
PyYAML==6.0
redis==4.4.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
twilio==7.16.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
vine==5.0.0
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
yarl==1.8.2
